Add functions and variables for the display
The variable for displaying the current operations and numbers is an
array for future search operations on the according button presses with
the operation symbols. It gets manipulated by pressing the number buttons and
operators by calling .push(). The according values are added to the string or removed by
the clear (C) button by calling .pop(). The AllClear button (AC)
empties the string array.